+++ framework/domain/src/main/java/is/codion/framework/domain/entity/EntityType.java
@@ -137,7 +137,7 @@ public interface EntityType {
 	 *             .caption("Active")
 	 *             .nullable(false)
 	 *             .defaultValue(true))
-	 *     .tableName("customer")
+	 *     .table("customer")
 	 *     .caption("Customer")
 	 *     .description("Customer information")
 	 *     .orderBy(ascending(Customer.NAME))